+++ stextools/snify/selection.py
@@ -9,9 +9,10 @@ from pylatexenc.latexwalker import LatexWalker, LatexMacroNode, LatexMathNode, L
 from stextools.core.linker import Linker
 from stextools.core.macros import STEX_CONTEXT_DB
 from stextools.core.simple_api import get_symbols, SimpleSymbol
-from stextools.snify.commands import Command, CommandInfo, CommandOutcome, SetNewCursor
+from stextools.stepper.command_outcome import CommandOutcome, SetNewCursor
+from stextools.stepper.commands import CommandInfo, Command
 from stextools.snify.skip_and_ignore import IgnoreList, SrSkipped
-from stextools.snify.state import SelectionCursor, State
+from stextools.stepper.state import SelectionCursor, State
 from stextools.snify.stemming import string_to_stemmed_word_sequence, string_to_stemmed_word_sequence_simplified
 from stextools.utils.linked_str import LinkedStr, string_to_lstr
 from stextools.utils.ui import standard_header